refactor(ItineraryChecker): Extract itinerary checking logic.
