+++ src/test/java/org/opentripplanner/middleware/tripmonitor/jobs/CheckMonitoredTripBasicTest.java
@@ -1,24 +1,18 @@
 package org.opentripplanner.middleware.tripmonitor.jobs;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.opentripplanner.middleware.itinerarymatching.ItineraryCheckStatus;
 import org.opentripplanner.middleware.models.ItineraryExistence;
 import org.opentripplanner.middleware.models.MonitoredTrip;
-import org.opentripplanner.middleware.otp.LegFinder;
 import org.opentripplanner.middleware.otp.OtpGraphQLVariables;
 import org.opentripplanner.middleware.otp.response.Itinerary;
-import org.opentripplanner.middleware.otp.response.Leg;
 import org.opentripplanner.middleware.tripmonitor.TripStatus;
 import org.opentripplanner.middleware.utils.DateTimeUtils;
 
 import java.time.DayOfWeek;
-import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
@@ -30,9 +24,6 @@ import java.util.Date;
 import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.opentripplanner.middleware.itinerarymatching.ItineraryMatchingUtils.createTransitWalkTransitItinerary;
-import static org.opentripplanner.middleware.utils.DateTimeUtils.convertToDate;
 import static org.opentripplanner.middleware.utils.DateTimeUtils.getOtpZoneId;
 
 /**
@@ -276,43 +267,5 @@ public class CheckMonitoredTripBasicTest {
                 "Don't send reminder for trip no longer possible."
             )
         );
-    }
-
-    @Test
-    void canComputeTripDelays() throws CloneNotSupportedException {
-        // Given a saved itinerary and some matching legs,
-        // CheckMonitoredTrip should be able to compute trip delays.
-        LocalDateTime baseTime = LocalDateTime.of(2025, 11, 10, 8, 0, 0);
-
-        Itinerary itinerary = createTransitWalkTransitItinerary(baseTime);
-        // Simulate a 6-minute delay on the itinerary arrival.
-        // For the first leg, make it a 4-minute delay on departure only.
-        final int DEPARTURE_DELAY_SECONDS = (int)Duration.ofMinutes(4).toSeconds();
-        final int FINAL_DELAY_SECONDS = (int)Duration.ofMinutes(6).toSeconds();
-        Itinerary mockItinerary = createTransitWalkTransitItinerary(baseTime.plusSeconds(FINAL_DELAY_SECONDS));
-        Leg firstMockLeg = mockItinerary.legs.get(0);
-        firstMockLeg.departureDelay = DEPARTURE_DELAY_SECONDS;
-        firstMockLeg.realTime = true;
-        firstMockLeg.startTime = convertToDate(LocalDateTime.ofInstant(firstMockLeg.startTime.toInstant(), getOtpZoneId()).minusMinutes(2));
-        Leg lastMockLeg = mockItinerary.legs.get(2);
-        lastMockLeg.arrivalDelay = FINAL_DELAY_SECONDS;
-        lastMockLeg.realTime = true;
-
-        MonitoredTrip trip = new MonitoredTrip();
-        trip.itinerary = itinerary;
-
-        MockLegResponseProvider mockLegResponseProvider = new MockLegResponseProvider(mockItinerary);
-        LegFinder mockLegFinder = new LegFinder(
-            mockLegResponseProvider::getLegResponse,
-            MockLegResponseProvider::computeLegIdForServiceDate
-        );
-
-        CheckMonitoredTrip check = new CheckMonitoredTrip(trip, mockLegFinder);
-        check.targetZonedDateTime = DateTimeUtils.nowAsZonedDateTime();
-        ItineraryCheckStatus itineraryCheckStatus = check.checkLegs();
-
-        assertTrue(itineraryCheckStatus.legsMatch);
-        assertEquals(DEPARTURE_DELAY_SECONDS, itineraryCheckStatus.departureDelaySeconds);
-        assertEquals(FINAL_DELAY_SECONDS, itineraryCheckStatus.arrivalDelaySeconds);
     }
 }
